Refactor: useHistory 삭제
HistoryContext로 합쳐서 사용
